+++ xbsl/rules/__init__.py
@@ -122,3 +122,11 @@ from . import (  # noqa: F401
     yaml_properties,
     yaml_types,
 )
+
+from . import ambiguous_types  # noqa: F401
+
+from . import unused_return_value  # noqa: F401
+
+from . import captured_local_write  # noqa: F401
+
+from . import missing_return  # noqa: F401